Create code
This code defines two functions, extractMineral() and updatePrice(), that update the mineralAmount and mineralPrice variables, respectively. It also uses the console.log() function to output messages indicating the current state of the mineral extraction and pricing.

You can use this as a basic template and add more functionality as per your requirement.